fix: refactor recycler view actions
