+++ app/src/espresso/java/xyz/lbres/trickcalculator/recyclerview/RecyclerViewMatcher.kt
@@ -1,10 +1,9 @@
-package xyz.lbres.trickcalculator.ui.attributions
+package xyz.lbres.trickcalculator.recyclerview
 
 import android.content.res.Resources
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import org.hamcrest.Description
-import org.hamcrest.Matcher
 import org.hamcrest.TypeSafeMatcher
 
 /**
